restrict some action's access to admin users

+++ routes/ptlas.js
@@ -1,12 +1,13 @@
 const express = require("express");
 const router = express.Router();
+const { authenticate } = require("../middlewares/auth");
 
 const { Ptla } = require("../models");
 
-router.get("/", async (req, res) => {
-  if (req.query.role) {
+router.get("/", authenticate, async (req, res) => {
+  if (req.query.position) {
     try {
-      const ptla = await Ptla.findOne({ role: req.query.role });
+      const ptla = await Ptla.findOne({ position: req.query.position });
       return res.status(200).json(ptla);
     } catch (err) {
       return res.status(500).json({ msg: "failed to find" });
@@ -28,17 +29,18 @@ router.get("/", async (req, res) => {
   }
 });
 
-// router.post("/", async (req, res) => {
-//   try {
-//     const newPtla = new Ptla({ ...req.body });
-//     const created = await newPtla.save();
-//     console.log(created);
-//     return res.status(201).json(created);
-//   } catch (err) {
-//     return res.status(500).json({ msg: "failed to create" });
-//   }
-// });
+router.get("/:id", authenticate, async (req, res) => {
+  try {
+    const { id } = req.params;
+    const ptla = await Ptla.findById(id);
+    return res.status(200).json(ptla);
+  } catch (err) {
+    return res.status(500).json({ msg: "failed to find" });
+  }
+});
 
+//  개발용
+// /*
 router.post("/", async (req, res) => {
   try {
     let count = 0;
@@ -54,19 +56,26 @@ router.post("/", async (req, res) => {
     return res.status(500).json({ msg: "failed to create" });
   }
 });
+// */
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", authenticate, async (req, res) => {
+  if (req.user.admin < 1) {
+    return res.status(403).json({ msg: "no privilege" });
+  }
   const { id } = req.params;
   try {
     await Ptla.findByIdAndUpdate(id, req.body);
-    const updated = await Ptla.findOne({ id });
+    const updated = await Ptla.findById(id);
     return res.status(200).json(updated);
   } catch (err) {
     return res.status(500).json({ msg: "failed to update" });
   }
 });
 
-router.delete("/", async (req, res) => {
+router.delete("/", authenticate, async (req, res) => {
+  if (req.user.admin < 2) {
+    return res.status(403).json({ msg: "no privilege" });
+  }
   try {
     const deleted = await Ptla.deleteMany({ admin: { $ne: 2 } }).exec();
     return res.status(200).json(deleted);
